moved schemas to models folder, tested registration, checkin, and export routes
