+++ client/scripts/factories/AdminService.js
@@ -5,12 +5,26 @@ myApp.factory('AdminService', ['$http', '$location', function($http, $location){
       $http.get('/reflection').then(function(response) {
         console.log('GOTTEN REFLECTIONS', response);
       });
+      getCSVforReflections();
   }
 
   function getRegistrationInfo() {
       $http.get('/register/registration').then(function(response) {
         console.log('GOTTEN REGISTRATIONS', response);
       });
+      getCSVforRegistration();
+  }
+
+  function getCSVforReflections(){
+    $http.get('/csvExport/reflections').then(function(response){
+      console.log('reflections in getCSV: ', response.data);
+    });
+  }
+
+  function getCSVforRegistration(){
+    $http.get('/csvExport/registration').then(function(response){
+      console.log('registrations in getCSV: ', response.data);
+    });
   }
 
 
@@ -19,7 +33,9 @@ myApp.factory('AdminService', ['$http', '$location', function($http, $location){
 return {
 
   getReflections: getReflections,
-  getRegistrationInfo: getRegistrationInfo
+  getRegistrationInfo: getRegistrationInfo,
+  getCSVforReflections : getCSVforReflections,
+  getCSVforRegistration : getCSVforRegistration
