+++ recoveree-reflections.csv
@@ -1,0 +1,2 @@
+"feelings.name","feelings.value","feelingsWhy","drugAlcoholIntake","medication","sleep","dream","whatDream","exercise","food","spnsrMntrConnect","groupMeet","commntyService","stressors.name","stressors.value","selfishDishonest","howSelfshDishnt","tomorrowGoal","dailyGoal","gratitude","peerSupport","counselor"
+,,,,,,,,,,,,,,,,,,,,,
