+++ server/routes/checkIn.js
@@ -2,47 +2,48 @@ var express = require('express');
 var router = express.Router();
 var passport = require('passport');
 var Users = require('../models/user');
+var CheckIn = require('../models/checkIn');
 var path = require('path');
 
 var mongoose = require("mongoose");
 
-var RecovreeSchema = mongoose.Schema({});
+// var RecovreeSchema = mongoose.Schema({});
 
-var CheckInSchema = mongoose.Schema({
-  feelings: {type: Array},
-  drugAlcoholIntake: {type: Boolean},
-  medication: {type: Boolean},
-  sleep: {type: Number},
-  dream: {type: Boolean},
-  exercise: {type: Number},
-  food: {type: Number},
-  spnsrMntrConnect: {type: Boolean},
-  groupMeet: {type: Number},
-  commntyService: {type: Boolean},
-  stressors: {type: Array},
-  selfishDishonest: {type: Boolean},
-  howSelfshDishnt: {type: String},
-  tomorrowGoal: {type: String},
-  dailyGoal: {type: String},
-  gratitude: {type: String},
-  peerSupport: {type: Boolean},
-  counselor: {type: Boolean},
-  checkInDate: {type: Date, default: Date.now},
-  // memberID: {type: Schema.ObjectId, ref: 'Registration'} //references Registration Schema
-});
+// var CheckInSchema = mongoose.Schema({
+//   feelings: {type: Array},
+//   drugAlcoholIntake: {type: Boolean},
+//   medication: {type: Boolean},
+//   sleep: {type: Number},
+//   dream: {type: Boolean},
+//   exercise: {type: Number},
+//   food: {type: Number},
+//   spnsrMntrConnect: {type: Boolean},
+//   groupMeet: {type: Number},
+//   commntyService: {type: Boolean},
+//   stressors: {type: Array},
+//   selfishDishonest: {type: Boolean},
+//   howSelfshDishnt: {type: String},
+//   tomorrowGoal: {type: String},
+//   dailyGoal: {type: String},
+//   gratitude: {type: String},
+//   peerSupport: {type: Boolean},
+//   counselor: {type: Boolean},
+//   checkInDate: {type: Date, default: Date.now},
+//   // memberID: {type: Schema.ObjectId, ref: 'Registration'} //references Registration Schema
+// });
+//
+// var Recovree = mongoose.model('recovree', RecovreeSchema, 'recovree');
+// var CheckIn = mongoose.model('checkIn', CheckInSchema, 'recovree');
 
-var Recovree = mongoose.model('recovree', RecovreeSchema, 'recovree');
-var CheckIn = mongoose.model('checkIn', CheckInSchema, 'recovree');
-
-router.get('/', function (req, res) {
-  Recovree.find({}, function(err, recovree){
-    if(err){
-      console.log("Mongo Error: ", err);
-      res.send(500);
-    }
-    res.send(listings);
-  });
-});
+// router.get('/', function (req, res) {
+//   Recovree.find({}, function(err, recovree){
+//     if(err){
+//       console.log("Mongo Error: ", err);
+//       res.send(500);
+//     }
+//     res.send(listings);
+//   });
+// });
 
 router.post("/", function(req,res){
   var reflection = req.body;
